Converted all floats to doubles
Fixed error in TAS_ramp example

+++ mcramp/LEGACY/geom/banana.py
@@ -9,10 +9,10 @@ import re
 
 class GBanana(GPrim):
     def __init__(self, radius=0, position=(0, 0, 0), height=0, mintheta=0, maxtheta=0, idx=0, ctx=None):
-        self.radius     = np.float32(radius)
-        self.height     = np.float32(height)
-        self.mintheta   = np.float32(mintheta)
-        self.maxtheta   = np.float32(maxtheta)
+        self.radius     = np.float64(radius)
+        self.height     = np.float64(height)
+        self.mintheta   = np.float64(mintheta)
+        self.maxtheta   = np.float64(maxtheta)
         self.position   = position
         self.idx        = idx
